% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_population.R
\docType{class}
\name{BasicPop}
\alias{BasicPop}
\title{Object-oriented implementation of a basic population}
\format{
\link{R6Class} object with fields and methods as documented below
}
\value{
\link{R6Class} object representing an agent-based model simulation
}
\description{
Object-oriented implementation of a basic population

Object-oriented implementation of a basic population
}
\section{Methods}{

The following methods are available:
\itemize{
\item \code{initialize()} the initialize method
\item \code{print()} the print method
}
}

\examples{
bp <- BasicPop$new()
bp


}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{Status}}{The current status of the simulation}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{Status}}{The current status of the simulation}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BasicPop$new()}}
\item \href{#method-clone}{\code{BasicPop$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasicPop$new(num_farms)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasicPop$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
