% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_spread.R
\docType{class}
\name{BasicSpread}
\alias{BasicSpread}
\title{Object-oriented implementation of a basic spread module}
\format{
\link{R6Class} object with fields and methods as documented below
}
\value{
\link{R6Class} object representing an agent-based model simulation
}
\description{
Object-oriented implementation of a basic spread module

Object-oriented implementation of a basic spread module
}
\section{Methods}{

The following methods are available:
\itemize{
\item \code{initialize()} the initialize method
\item \code{print()} the print method
}
}

\examples{
bp <- BasicPop$new(50); bp
bp$set_extbeta(c(1, rep(0,49)))
bp$update(1)
bp$set_extbeta(c(0, rep(0,49)))
bp$update(99)
bp$compartments
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{Status}}{The current status of the simulation}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{Status}}{The current status of the simulation}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BasicSpread$new()}}
\item \href{#method-update}{\code{BasicSpread$update()}}
\item \href{#method-clone}{\code{BasicSpread$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasicSpread$new(population, probability = 0.1, proportion = 0.05)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update"></a>}}
\if{latex}{\out{\hypertarget{method-update}{}}}
\subsection{Method \code{update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasicSpread$update()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasicSpread$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
