% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blofeld-package.R
\docType{package}
\name{Blofeld-package}
\alias{Blofeld}
\alias{Blofeld-package}
\title{Blofeld: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Matt Denwood \email{md@sund.ku.dk}

Authors:
\itemize{
  \item Mossa Merhi \email{mossa@sund.ku.dk}
}

}
\keyword{internal}
