% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blofeld-package.R
\docType{package}
\name{blofeld-package}
\alias{blofeld}
\alias{blofeld-package}
\title{blofeld: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Matt Denwood \email{md@sund.ku.dk}

}
\keyword{internal}
